+++ oidc_provider/admin.py
@@ -4,6 +4,7 @@ from uuid import uuid4
 
 from django.forms import ModelForm
 from django.contrib import admin
+from django.utils.translation import ugettext_lazy as _
 
 from oidc_provider.models import Client, Code, Token, RSAKey
 
@@ -48,6 +49,17 @@ class ClientForm(ModelForm):
 @admin.register(Client)
 class ClientAdmin(admin.ModelAdmin):
 
+    fieldsets = [
+        [_(u''), {
+            'fields': ('name', 'client_type', 'response_type','_redirect_uris', 'jwt_alg'),
+        }],
+        [_(u'Credentials'), {
+            'fields': ('client_id', 'client_secret'),
+        }],
+        [_(u'Information'), {
+            'fields': ('contact_email', 'website_url', 'terms_url', 'logo', 'date_created'),
+        }],
+    ]
     form = ClientForm
     list_display = ['name', 'client_id', 'response_type', 'date_created']
     readonly_fields = ['date_created']