+++ oidc_provider/lib/utils/params.py
@@ -1,7 +1,0 @@
-
-
-class Params(object):
-    """
-    The purpose of this class is for accesing params via dot notation.
-    """
-    pass